+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update layout of current and cursor color [#40](https://github.com/vv9k/epick/pull/40)
 - Zoom window can now be enabled by dragging the zoom button, this way the main window stays in focus and colors from other windows can be extracted. [#41](https://github.com/vv9k/epick/pull/41)
 - Set app icon on OS taskbar
+- Revert back to default display format when currently used custom format is deleted
 
 # 0.6.0
 - Add support for persistent configuration. A configuration file with a YAML syntax will be loaded on startup from appropriate config directory depending on OS.